src: make sure the job runs on the exact date
It can happen that a recurrent job is scheduled a little before (less
than a second) the expected date. If this happens, the next recurrence
happens in the expected data (less than a second later) so a job is run
twice at the same time. To avoid this, don't run the job when initially
scheduled and reschedule the job to run in the expected date.

PR-URL: https://github.com/node-schedule/node-schedule/pull/396
Fixes: https://github.com/node-schedule/node-schedule/issues/263
Fixes: https://github.com/node-schedule/node-schedule/issues/370