doc: fix cron string diagram consistency
The vertical line character used in the weekday field is different than
the vertical line characters used for the other fields; it is a pipe
character (EASCII 124) rather than a vertical line (EASCII 179).  All
other line characters in the diagram are from the line-drawing
character set, so I believe the intention was to use such a character
in this case as well.

PR-URL: https://github.com/node-schedule/node-schedule/pull/404