Update README.md
The example with chron format and a timezone was broken. This PR simply fixes README.md.

Explanation of README.md change:
If you provide 2 strings, the first one gets interpreted as the name of the job and the second one gets interpreted as the chron schedule. 3 strings must be provided in order to set the timezone: job name, chron schedule, and timezone.