Emit an `error` event when invoking a job causes an error
Previously, if a job was invoked and caused an error, it would cause an
unhandled exception or an unhandled promise rejection. The only way for
a user to handle it was to wrap the actual job in a wrapper function.

This makes it easier to handle by catching the error and emitting an
`error` event when an error is thrown synchronously or a rejected
Promise is a returned.
As a bonus, the result of the job is returned when `job.invoke()` is
called.