Emit a success event after a successful job invocation.
When the callback of a job invocation returns successfully (i.e. returning
without raising an error), a `success` event is emitted receiving the data
returned by the callback. In case of the callback returning a promise, the
`success` event is emitted if and only if the promise resolves successfully,
receiving the resolved value.